Refactoring: GeoDataRiverSurvey::AddExtensionCommand is moved to separated files.
