+++ libs/geodata/riversurvey/riversurvey.pro
@@ -144,6 +144,7 @@ HEADERS += gd_riversurvey_global.h \
            geodatariversurveyproxy.h \
            geodatariversurveyvtkexporter.h \
            gridcreatingconditionriversurveyinterface.h \
+           private/geodatariversurvey_addextensioncommand.h \
            private/geodatariversurvey_setdisplaysettingcommand.h \
            private/geodatariversurveycrosssectionwindow_datatabledelegate.h \
            private/geodatariversurveycrosssectionwindow_impl.h \
@@ -189,6 +190,7 @@ SOURCES += geodatarivercrosssection.cpp \
            geodatariversurveynodemapper.cpp \
            geodatariversurveyproxy.cpp \
            geodatariversurveyvtkexporter.cpp \
+           private/geodatariversurvey_addextensioncommand.cpp \
            private/geodatariversurvey_setdisplaysettingcommand.cpp \
            private/geodatariversurveycrosssectionwindow_datatabledelegate.cpp \
            private/geodatariversurveycrosssectionwindow_riversurveytabledelegate.cpp \
